Point the screens at the new RPCs, and stop deriving money on the phone
The server-side half of this branch moved the arithmetic into the
database. This is the client catching up: the collection, settlement,
transfer and draft paths now call the RPCs and render what they return,
instead of computing a classification or a closing balance locally and
hoping it agreed with the server.

Collection Mode handles the duplicate_warning answer - another member
already collected this loan today - by offering Close or Continue, and
re-calls with p_confirm_duplicate on Continue. The loan wizard handles
INSUFFICIENT_FLOAT by saving a draft rather than surfacing a raw
database error. Today's Route and the due lists read v_collection_due,
so installmentDue and the overdue flag are the server's numbers now,
not the three placeholder values the client used to substitute.

network_error_handler learns 429, which the auth rate limiter returns.

Also fixes three things that stopped this tree compiling at all - it had
never been through flutter analyze:

  - lr_009 imported local_auth_flutter_biocredentials, a package that
    does not exist on pub.dev and was never in pubspec.yaml. Unused;
    removed, with the equally unused local_auth import.
  - design_showcase passed compact: to ManaLanguageSelector, a parameter
    dropped when that widget was rewritten.
  - an unused import in lr_002.

flutter analyze is clean and all 175 tests pass.

+++ lib/features/investor_workspace/screens/iw_004_request_withdrawal.dart
@@ -111,8 +111,8 @@ class _WithdrawalFormState extends ConsumerState<_WithdrawalForm> {
     super.dispose();
   }
 
-  double? get _requestedAmount =>
-      double.tryParse(_amountController.text.trim());
+  int? get _requestedAmount =>
+      int.tryParse(_amountController.text.trim());
 
   bool get _canSubmit {
     if (widget.disabled) return false;
